[refactor] custom interface 구조 정리
