+++ src/components/Home/Confirm/ConfirmWrite/ConfirmWriteHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { ReactComponent as ConfirmWriteCloseSVG } from '../../../../styles/images/icons/confirm-comments-close.svg';
-import { IConfirmWriteHeaderProps } from 'types/Home';
+import { IConfirmWriteHeaderProps } from 'types/Home/Confirm';
 
 function ConfirmWriteHeader({ onSubmit }: IConfirmWriteHeaderProps) {
   return (
