+++ src/components/Home/Confirm/ConfirmWrite/ConfirmWriteImageSlideInput.tsx
@@ -4,7 +4,8 @@ import 'swiper/css/pagination';
 import { Pagination } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { ReactComponent as ConfirmWriteSliderSVG } from '../../../../styles/images/icons/confirm-write__slider.svg';
-import { IConfirmWriteImageSlideInputProps, IImageFile } from 'types/Home';
+import { IConfirmWriteImageSlideInputProps } from 'types/Home/Confirm';
+import { IImageFile } from 'types/Common';
 
 function ConfirmWriteImageSlideInput({
   setInputImages,
