+++ src/page/confirm/ConfirmWrite.tsx
@@ -1,12 +1,13 @@
 import React, { useState } from 'react';
 import ConfirmWriteHeader from 'components/Home/Confirm/ConfirmWrite/ConfirmWriteHeader';
 import { registerConfirm } from 'api/confirm';
-import { IImageFile, IRegisterApiResponse } from 'types/Home';
 import { useApi } from 'hooks/useApi';
 import useApiNavigation from 'hooks/useApiNavigation';
 import ConfirmWriteTextInput from 'components/Home/Confirm/ConfirmWrite/ConfirmWriteTextInput';
 import ConfirmWriteVote from 'components/Home/Confirm/ConfirmWrite/ConfirmWriteVote';
 import ConfirmWriteImageSlideInput from 'components/Home/Confirm/ConfirmWrite/ConfirmWriteImageSlideInput';
+import { IRegisterApiResponse } from 'types/Home/Confirm';
+import { IImageFile } from 'types/Common';
 
 function ConfirmWrite() {
   const apiNavigation = useApiNavigation<IRegisterApiResponse>();
